Created util file for schedules
