+++ src/leo-server-app/src/utils/satellite.utils.ts
@@ -107,4 +107,6 @@ export const getNextPasses = (noradId: string) => {
       }
     }
   }
+
+  return nextPasses;
 };
